Started on property access check commands with string length
There are a lot of commands that may come out of a property access expression.

Furthermore, it should be noted that `let length = "abc".length` would normally become a float instead of an int because TypeScript doesn't know the difference. To deal with hardcodings such as length commands, a concept of type adjustments is added. The variable declaration has an added third step of typing changes: if a value is received, if there is a potential type adjustment, it's taken. Length commands are the first known adjustment in this line.